+++ lib/sensu-plugins-mongodb/metrics.rb
@@ -377,7 +377,9 @@ module SensuPluginsMongoDB
         address_str = "#{host}:#{port}"
         client_opts = {}
         client_opts[:database] = db_name
-        client_opts[:connect] = connect
+        unless connect.nil?
+          client_opts[:connect] = connect
+        end
         unless db_user.nil?
           client_opts[:user] = db_user
           client_opts[:password] = db_password